Add default audio file to play at startup.
All Bateleur code is designed such that there must be an audio file
loaded at all times. To deal with this on the first startup (when the
library has not yet been loaded) a default 'Welcome to Bateleur' audio
file has been created.
In future, this audio file could also be used to play some audio telling
the user how to set up the application.